Get rid of program_options from headers, and limited them in common/command_line_functions.h and tests/core_tests/chaingen_program_options.h
Where there was the original options header, there's fwd now
