+++ src/CMakeLists.txt
@@ -83,6 +83,7 @@ endfunction ()
 include(Version)
 monero_add_library(version SOURCES ${CMAKE_BINARY_DIR}/version.cpp DEPENDS genversion)
 
+add_subdirectory(fwd)
 add_subdirectory(common)
 add_subdirectory(crypto)
 add_subdirectory(ringct)
