+++ src/blockchain_utilities/blockchain_depth.cpp
@@ -29,7 +29,7 @@
 #include <boost/range/adaptor/transformed.hpp>
 #include <boost/filesystem/path.hpp>
 #include <boost/algorithm/string.hpp>
-#include "common/command_line.h"
+#include "common/command_line_functions.h"
 #include "common/varint.h"
 #include "cryptonote_core/tx_pool.h"
 #include "cryptonote_core/cryptonote_core.h"
