+++ src/daemonizer/daemonizer.h
@@ -28,9 +28,8 @@
 
 #pragma once
 
+#include "fwd/boost_monero_program_options_fwd.h"
 #include <boost/filesystem/path.hpp>
-#include <boost/program_options/options_description.hpp>
-#include <boost/program_options/variables_map.hpp>
 
 namespace daemonizer
 {
