+++ src/fwd/CMakeLists.txt
@@ -26,74 +26,29 @@
 # STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
 # THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-set(daemon_sources
-  command_parser_executor.cpp
-  command_server.cpp
-  daemon.cpp
-  executor.cpp
-  main.cpp
-  rpc_command_executor.cpp
+set(TGT fwd)
+set(fwd_sources
+  monero_fwd.cpp
 )
 
-set(daemon_headers)
+set(fwd_headers
+  monero_fwd.h
+  boost_monero_fwd.h
+  boost_monero_program_options_fwd.h
+)
 
-set(daemon_private_headers
-  command_parser_executor.h
-  command_server.h
-  command_line_args.h
-  core.h
-  daemon.h
-  executor.h
-  p2p.h
-  protocol.h
-  rpc.h
-  rpc_command_executor.h
+set(fwd_private_headers
+)
 
-  # cryptonote_protocol
-  ../cryptonote_protocol/cryptonote_protocol_defs.h
-  ../cryptonote_protocol/cryptonote_protocol_handler.h
-  ../cryptonote_protocol/cryptonote_protocol_handler.inl
-  ../cryptonote_protocol/cryptonote_protocol_handler_common.h
+monero_private_headers(${TGT}
+  ${fwd_private_headers})
+  
+monero_add_library(${TGT}
+  ${fwd_sources}
+  ${fwd_headers}
+  ${fwd_private_headers}
+)
 
-  # p2p
-  ../p2p/net_node.h
-  ../p2p/net_node_common.h
-  ../p2p/net_peerlist.h
-  ../p2p/net_peerlist_boost_serialization.h
-  ../p2p/p2p_protocol_defs.h
-  ../p2p/stdafx.h)
+#monero_install_headers(common
+#  ${fwd_headers})
 
-monero_private_headers(daemon
-  ${daemon_private_headers})
-monero_add_executable(daemon
-  ${daemon_sources}
-  ${daemon_headers}
-  ${daemon_private_headers})
-target_link_libraries(daemon
-  PRIVATE
-    rpc
-    blockchain_db
-    cryptonote_core
-    cncrypto
-    common
-    p2p
-    cryptonote_protocol
-    daemonizer
-    serialization
-    daemon_rpc_server
-    ${EPEE_READLINE}
-    version
-    ${Boost_CHRONO_LIBRARY}
-    ${Boost_FILESYSTEM_LIBRARY}
-    ${Boost_PROGRAM_OPTIONS_LIBRARY}
-    ${Boost_REGEX_LIBRARY}
-    ${Boost_SYSTEM_LIBRARY}
-    ${CMAKE_THREAD_LIBS_INIT}
-    ${ZMQ_LIB}
-    ${GNU_READLINE_LIBRARY}
-    ${EXTRA_LIBRARIES}
-    ${Blocks})
-set_property(TARGET daemon
-  PROPERTY
-    OUTPUT_NAME "monerod")
-install(TARGETS daemon DESTINATION bin)
