+++ src/fwd/monero_fwd.h
@@ -27,33 +27,12 @@
 // THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #pragma once
-#include "fwd/boost_monero_program_options_fwd.h"
 
-#undef MONERO_DEFAULT_LOG_CATEGORY
-#define MONERO_DEFAULT_LOG_CATEGORY "daemon"
+/*
+This header could be included globally, as long as it's kept clean of any normal declarations.
+However it would help in better understanding the structure of dependencies,
+if each fwd was still included individually on demand.
+*/
 
-namespace daemonize {
-
-struct t_internals;
-
-class t_daemon final {
-public:
-  static void init_options(boost::program_options::options_description & option_spec);
-private:
-  void stop_p2p();
-private:
-  std::unique_ptr<t_internals> mp_internals;
-  uint16_t public_rpc_port;
-public:
-  t_daemon(
-      boost::program_options::variables_map const & vm,
-      uint16_t public_rpc_port = 0
-    );
-  t_daemon(t_daemon && other);
-  t_daemon & operator=(t_daemon && other);
-  ~t_daemon();
-
-  bool run(bool interactive = false);
-  void stop();
-};
-}
+// Include here other monero forward declaration aggregates
+#include "fwd/boost_monero_fwd.h"
