+++ src/simplewallet/simplewallet.h
@@ -38,8 +38,7 @@
 #include <memory>
 
 #include <boost/optional/optional.hpp>
-#include <boost/program_options/variables_map.hpp>
-
+#include "fwd/boost_monero_program_options_fwd.h"
 #include "cryptonote_basic/account.h"
 #include "cryptonote_basic/cryptonote_basic_impl.h"
 #include "wallet/wallet2.h"
