+++ src/wallet/wallet_args.cpp
@@ -32,6 +32,7 @@
 #include <boost/format.hpp>
 #include "common/i18n.h"
 #include "common/util.h"
+#include "common/command_line_functions.h"
 #include "misc_log_ex.h"
 #include "string_tools.h"
 #include "version.h"
