+++ src/wallet/wallet_rpc_server.cpp
@@ -39,7 +39,7 @@ using namespace epee;
 #include "version.h"
 #include "wallet_rpc_server.h"
 #include "wallet/wallet_args.h"
-#include "common/command_line.h"
+#include "common/command_line_functions.h"
 #include "common/i18n.h"
 #include "common/scoped_message_writer.h"
 #include "cryptonote_config.h"
