+++ tests/core_tests/chaingen_main.cpp
@@ -28,7 +28,9 @@
 // 
 // Parts of this file are originally copyright (c) 2012-2013 The Cryptonote developers
 
+#include "common/command_line_functions.h"
 #include "chaingen.h"
+#include "chaingen_program_options.h"
 #include "chaingen_tests_list.h"
 #include "common/util.h"
 #include "common/command_line.h"
